rbtree: API changes. Replace is_unique with is_stable.
This patch changes the red-black tree to always allow duplicate keys. We still
need to determine if a tree sorts the duplicate nodes in a stable order
(ensuring FIFO ordering for find operations). So the is_unique field has been
replaced with is_stable which is symmetrically the opposite of is_unique:
if is_unique was true (false) then is_stable should be false (true).
Maintaining stability in a red-black tree imposes extra costs when searching
so if stability is not required slightly better performance can be obtained.