rbtree: Add finger versions of Insert and Find.
A finger is a pointer to a node in a tree and is used like a hint for
algorithms that traverse the tree. If the hint is bad performance may
decrease although the asymptotic complexity of the algorithms are not
changed. Using the tree's root as a finger yields the default behavior
of these algorithms.

This patch also cleans up Insert and Find, and fixes a bug in Insert
that left interrupts disabled after returning.