rbtree: Replace the color field of RBTree_Node with opaque attributes field
Using an enum type to store the color of a node will use more space than
strictly required. Instead we will store the color as a single bit in an
opaque attributes field that (for now) is 8 bits long. The attributes field
can be used to store additional metadata about red-black tree nodes.